+++ pydantic_ai_harness/experimental/step_persistence/__init__.py
@@ -1,19 +1,20 @@
 """Step-event persistence: append-only event log, continuable snapshots, tool-effect ledger."""
 
-from pydantic_ai_harness.step_persistence._capability import StepPersistence
-from pydantic_ai_harness.step_persistence._helpers import (
+from pydantic_ai_harness.experimental._warn import warn_experimental
+from pydantic_ai_harness.experimental.step_persistence._capability import StepPersistence
+from pydantic_ai_harness.experimental.step_persistence._helpers import (
     annotate_tool_effect,
     continue_run,
     fork_run,
     is_provider_valid,
 )
-from pydantic_ai_harness.step_persistence._store import (
+from pydantic_ai_harness.experimental.step_persistence._store import (
     FileStepStore,
     InMemoryStepStore,
     SqliteStepStore,
     StepStore,
 )
-from pydantic_ai_harness.step_persistence._types import (
+from pydantic_ai_harness.experimental.step_persistence._types import (
     ContinuableSnapshot,
     EventKind,
     RunRecord,
@@ -21,6 +22,8 @@ from pydantic_ai_harness.step_persistence._types import (
     ToolEffectRecord,
     ToolEffectStatus,
 )
+
+warn_experimental('step_persistence')
 
 __all__ = [
     'ContinuableSnapshot',
