+++ pydantic_ai_harness/experimental/step_persistence/_helpers.py
@@ -13,9 +13,9 @@ from pydantic_ai.messages import (
 )
 from pydantic_ai.tools import RunContext
 
-from pydantic_ai_harness.step_persistence._context import current_run_id
-from pydantic_ai_harness.step_persistence._store import StepStore
-from pydantic_ai_harness.step_persistence._types import ToolEffectRecord
+from pydantic_ai_harness.experimental.step_persistence._context import current_run_id
+from pydantic_ai_harness.experimental.step_persistence._store import StepStore
+from pydantic_ai_harness.experimental.step_persistence._types import ToolEffectRecord
 
 
 def is_provider_valid(messages: list[ModelMessage]) -> bool:
