+++ tests/experimental/media/cassettes/test_s3_cassettes/TestS3MediaStoreCassettes.test_get_round_trips_bytes.yaml
