+++ tests/experimental/media/cassettes/test_s3_cassettes/TestS3MediaStoreCassettes.test_put_succeeds.yaml
